backend/requirements.txt
altgraph==0.17.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
ecdsa==0.19.1
fastapi==0.121.3
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
hid==1.0.8
hidapi==0.14.0.post4
httptools==0.7.1
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.3.3
passlib==1.7.4
pefile==2023.2.7
pillow==11.2.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32-ctypes==0.2.3
pywinusb==0.4.2
PyYAML==6.0.3
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
